Backend/requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==25.1.0
asgiref==3.9.1
attrs==25.3.0
beautifulsoup4==4.12.2
bleach==6.1.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.2
colorama==0.4.6
coverage==7.3.2
cryptography==41.0.7
dj-database-url==2.1.0
Django==4.2.7
django-cors-headers==4.3.1
django-extensions==3.2.3
django-filter==23.1
django-guardian==2.4.0
django-ratelimit==4.1.0
django-redis==5.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.26.5
drf-yasg==1.21.7
factory-boy==3.3.0
Faker==20.1.0
gunicorn==21.2.0
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
lxml==4.9.3
packaging==25.0
Pillow==10.1.0
pluggy==1.6.0
psutil==6.1.0
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.8.0
pytest==7.4.3
pytest-django==4.7.0
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
pytz==2025.2
PyYAML==6.0.2
redis==5.0.1
referencing==0.36.2
requests==2.31.0
rpds-py==0.27.0
sentry-sdk==1.38.0
six==1.17.0
soupsieve==2.7
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
webencodings==0.5.1
whitenoise==6.6.0
